Fix error handling in paging query options
Prior to this commit, there was a bug that was causing the
'order-by' query parameter to be processed incorrectly.

The root cause was that a bunch of the paging utility functions
were trying to return ring error responses, rather than
raising an exception and letting the actual ring middleware
deal with the ring side of things.

This was really, really dumb, and this commit fixes it.