Refactor resource-identity-hash tests to be more relevant
This patch decomposes the resource-identity-hash so that the serialize
mechanism is separate and test-able without the memoize front-end.

It also expands the testing for sort-nested-maps to cover other data
types, does more analysis of the parts to make sure they are sorted as
well as apposed to internal comparison as clojure will always make
this true regardless of sort order.

Signed-off-by: Ken Barber <ken@bob.sh>