(maint) Lift variables out of conditionals
These variables (initscriptname, log_dir) are always composed of a string
followed by the @name variable, so it makes little sense to repeat their
definition in each branch of the if @pe conditional. This commit moves them
below the conditional, where @name is defined and they can be safely defined.