Log errors when invoking a subcommand
The main dispatcher scans available namespaces and dispatches to the
specified "-main" function. For example, running the "services"
subcommand would trigger loading of the "-main" function in the
"com.puppetlabs.puppetdb.cli.services" namespace.

However, if there was an error during the attempt to "require" that
namespace, it wouldn't be logged properly. This commit adds the
"require" step to the existing try/catch block.