Use simplified `order-by` expression data structure
Prior to this commit, we were using two different data structures
to represent an 'order-by' term.

We also weren't validating the ascending/descending portion of
an order-by term in the HTTP middleware; we were putting that
off all the way until the jdbc code, and passing around these
potentially varying string representations everywhere else
throughout the code.

This commit moves the validation logic for the 'order' terms back
into the HTTP middleware.  From that point on, we can be assured
that the 'order' will be represented as either `:ascending` or
`:descending`, and all of the order-by terms will be valid
based on the `order-by-expr?` predicate that goes along with
the new `order-by` utility function.