Stringify the facts in the facts terminus test
Prior to this commit, when we were testing the pson serialization
of facts, we were not calling 'stringify' first; in the actual
terminus code, we *do* call stringify.

A recent change in the `master` branch of Puppet started causing
this test to fail due to subtle differences in the pson code path
based on whether you have called stringify or not.

This commit simply makes the test consistent with the actual
terminus code.