(#22882) Use v3 resources/facts queryies for /v3/nodes/*/resources|facts
Currently the /v3/nodes/*/resources|facts end-point uses the V2 version
internally, this patch ammends the end-point to use the v3 end-point.

In particular the sourcefile/sourceline attributes have been changed to
file line which makes the resources sub-end-point inconsistent with its
root-level equivalent. It also means the returned data is not compatible
with the catalog wireformat.

Signed-off-by: Ken Barber <ken@bob.sh>