Move `explore` tests to v2 and v3 namespaces
Prior to this commit, the `explore` http tests were under the
`experimental` namespace.  This commit simply moves them
to v2 and creates a copy of them for v3.

As of this commit, the tests pass in both places; this is a bug.
I'm saving the bugfix for a second commit in order to make the
diff more readable.