Fix bug in paging test utility code
At some point when we changed the API for paging to use
`include-total` instead of `count?` for indicating that
we'd like for the response to include a count header,
we missed a spot in the test utility code that was still
expecting it to be `count?`, and thus we lost test
coverage for the count-related functionality.

This commit fixes that oversight and adds more strict
validation to the arguments to the test utility function,
to ensure that it is being used properly by all callers.