(packaging) (#17555) Fix issues with systemctl path in the RPM spec file
In the previous commit for this issue, the path to the systemctl utility
was incorrect in the %preun and %postun sections. Several calls were pointing
to /usr/sbin/systemctl instead of the correct path used in the rest of the
spec, /usr/bin/systemctl. This would lead to an RPM that would give errors
on uninstall.

This commit corrects the path to the systemctl utility.