(#22960) Don't automatically sort fact query results
Previously, we'd sort fact query results by default, regardless of
whether or not the user has requested sorting. That incurs a performance
penalty, as the DB has to now to a costly sort operation. This patch
removes the sort, and if users want sorted results they can use the new
sorting parameters to ask for that explicitly.