Run gem install of activerecord with greater verbosity
We keep getting activerecord gem installation problems, this patch changes
the verbosity of that installation so we can possibly see what is going
wrong. As this is intermittent its hard to test it in an isolated way, so
leaving this modification in source until it happens again seems a reasonable
approach.

Signed-off-by: Ken Barber <ken@bob.sh>