(PDB-81) - Adding a JDK 1.6 deprecation warning message
Adding a warning message to the console and the logs.  The bootstrapped
default logger is logged when the user provided log config is not there
(such as invoking the version command). It's then logged in services.clj
as well (after the logging config would have been applied).

Also made some some changes to c.p.p.core.clj to allow testing of the
main launch function (and the deprecation messages).