Add environment variable to allow SNAPSHOT dependencies
By default, leinengen doesn't allow SNAPSHOT dependencies when
running `uberjar` on a project that is not also a SNAPSHOT.
Because of the way that we handle the project version in the
PuppetDB project.clj file, PuppetDB is *never* in a SNAPSHOT
state.

This commit adds an environment variable that will cause
leinengen to allow SNAPSHOT deps, but only when running
acceptance tests FROM SOURCE.  Packaging jobs should still
fail if this kind of scenario is encountered for now.