Remove try/catch for with-transaction and use only wire-format for testing catalog concurrency
This patch removes the try/catch as a 'premature' idea for now, in favour of
doing this at presentation level (in commands perhaps) so we can capture both
exceptions as its been noticed the combination of exceptions is often required
versus just one.

It also fixes tests for concurrency updates, by using the wire-catalog only
from examples (we were using the non-wire one in their also but turns out they
were not exactly the same). And now that we no longer get the inner exception
with the try/catch removal above, the regexp needed ammending.

Signed-off-by: Ken Barber <ken@bob.sh>