(PDB-254) Change benchmark to mutate catalog resources and edges send a defined number of messages
Previously benchmark.clj when it would mutate a catalog would only
add a single resource.  This change will add a new resource or
mutate a random existing resource.  It will also add a new or
change an existing edge.  One of these four mutations is picked
at random.  This commit also adds a new parameter to benchmark.clj
to allow the syncronous sending of a specified number of commands
(per host) via the -N argument