PDB-249 Fix command-processing configuration exceptions
The use of (s/maybe s/Int) for the configuration options store-usage and
temp-usage were throwing exceptions when these configuration options were
used.

So this is really a bug in the convert-schema support, but in our case
we don't need to support the concept of 'nil' or 's/Int' for these items
(which is what s/maybe effectively does) since they are already marked as
optional.

This patch removes the use of s/maybe removing the exception, but does not
fix the original bug.

Signed-off-by: Ken Barber <ken@bob.sh>