PDB-307 Fix private method rspec failures due to mocha 1.0.0 upgrade
This patch fixes an rspec failure we were having on the terminus due to the
recent bump from mocha from 0.14.1 to 1.0.0.

Previously mocha seems to have been ignoring private method calls for whatever
reason. With the new version of mocha it is throwing exceptions (as it rightly
should).

This patch removes the use of the `private` declaration and converts all of
these methods to be document `@api private` in YARD instead migrating a
language enforcement to a documentation enforcement instead.

As an aside I noticed that char_encoding.rb had a 3 space indent convention,
this has been converted to 2 space as per our standard.

Signed-off-by: Ken Barber <ken@bob.sh>

Conflicts:
	Gemfile