PDB-307 Fix ruby-1.8.5 support for using bundler during rspec tests
Ruby 1.8.5 has some limitations in what it supports as far as gems go and
since we have been attempting as part of PDB-307 to use bundler more often
we have hit some problems.

* nokogiri, the version that beaker expects is not supported
* rake > 0.8.7 is also not supported

This patch removes beaker as a dependency while we run the rspec tests. This
will avoid pulling in lots of extra gems and hopefully speed up the whole
thing anyway. I've also applied this to the travis tests to hopefully speed
those up as well.

There is also now a conditional to change the dependencies for `rake` when
applied to a 1.8.5 ruby.

Signed-off-by: Ken Barber <ken@bob.sh>