(PDB-313) Fix catalog resource metadata not updating with parameters change
Catalog resource metadata fields like line number, file and exported were
not being updated properly when only one of those fields changed. Tags have
a similar problem, but the bug is not likely to occur in a real environment
(see PDB-332).

This commit issues an update that includes each updated catalog resource
metadata column when a difference in that column's value is found.