PDB-306 Switch to using new el5 imgages to fix DNS resolution hassles
This patch switches us to use Ashley Penney's new el5 images.

The problem we were seeing was that dhclient was seeing eth1 up at start up
and trying to use that for its authoritative DHCP interface, thus the box
was not getting correct DNS in /etc/resolv.conf (it was just falling back
to the baked in defaults).

This came down to 'kudzu' trying to add an interface automatically in
/etc/sysconfig/network-scripts/ifcfg-eth1 with dhcp specific details
however, vagrant won't remove this existing configuration when adding a
private network - it will just append.

Removing kudzu from the image did the trick. Now vagrant appending its
configuration will create the desired result.

Signed-off-by: Ken Barber <ken@bob.sh>