Don't use global schema validation
Prior to this commit, the `puppetdb.schema` namespace contained
a top-level call to primsatic's `set-fn-validation!` fn.  This
enabled strict schema validation for every occurence of a schema
annotation anywhere inside the JVM.

This behavior means that other libraries that use schemas lose
control over when strict validation occurs and when it does not,
which can cause unexpected behavior and potentially even bugs.

This commit removes the global toggle and instead prefers
explicit metadata to enforce validation on the various functions
that require it.