(packaging) correct path to libexec files for PE
Prior to this commit, when we installed to the libexec dir when building PE, we
installed to /opt/puppet/libexec/puppetdb. However, the %{name} macro in the
redhat spec will evaluate to pe-puppetdb. This means the %files section looked
for files that don't exist. This commit updates the spec file to use the
  correct %{realname} macro, e.g. 'puppetdb', which is the correct path.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>