(PDB-425) Move daemon-reload out of install-only
On upgrades, systemd service files may not have already been present, and they
might have changed. To address this, this commit moves the daemon-reload call
to systemctl out of the install only section of the %post, so that it is always
invoked. This will register the systemd service on upgrade from a sysv service
(as happens on fedora 18 through fedora 20) as well as register any changes to
the service file.