(PDB-425) Try SysV and SystemD on upgrades
On upgrades from PuppetDB < 1.6.0, even on platforms that support systemd, only
sysv will be in place. This means that when the status is queried in the %pre,
there won't be a systemd service file yet, so puppetdb will appear to not be
running even if it is. Next, when the service is stopped, also in the pre,
again there will be no systemd service to stop, even if there is a sysv service
running. This commit addresses that by testing for the existence of the systemd
service file, and if that is present, use the systemd service calls. Otherwise
it calls out to the sysv init system.