PDB-407 Fix bug in init script for fedora
While working on acceptance tests we've found a bug with the init script for
Fedora and 1.5.x specifically. That is, the ps that was being returned during
find_my_pid did not return the fully qualified file path to be matched
against. This patch simply takes the basename from JAVA_BIN and uses that
in the match instead.

Signed-off-by: Ken Barber <ken@bob.sh>