(PDB-468) Update scripts to use correct JAVA_BIN
The systemd service file and some of the other utility scripts were
using a hard-coded path to /usr/bin/java, instead of the java path
set in the system defaults file.

This commit updates those scripts to source the defaults file where
available, and fall back to sane defaults otherwise.