Use v3 end-point for import and benchmark tools
The function 'submit-command-via-http!' was still posting commands to the
/v2/command end-point which is no longer the latest. This wouldn't have
been a major problem, but it might have messed up some of the metrics based
http stats for users.

The patch switches the function to use the v3 endpoint, in this case an
unversioned end-point would be perfect but we lack that functionality.

Signed-off-by: Ken Barber <ken@bob.sh>