Add `modify_config_setting` helper method for acceptance
Prior to this commit, a lot of the acceptance tests were
written to just concatenate new config settings to the
end of the config files.  This relied on the old config
behavior where we allowed multiple occurrences of the
same config setting, and the last occurrence would win.

This commit adds a helper method that uses the inifile
module to modify the settings in place, rather than
adding duplicate copies of them.  It also updates
all of the tests that were written using the concatenation
pattern to use the new method instead.