PDB-470 Provide new db setting 'statements-cache-size' with a default of 1000
This setting adjusts how many SQL prepared statements get cached via BoneCP.

By using this setting we've seen an almost 40% decrease in wall clock time it
takes to store a new catalog.

This patch adds the new configuration item as a user configurable one, with a
default set to 1000 for now. Documentation has also been added for this
setting.

Signed-off-by: Ken Barber <ken@bob.sh>