(maint) Replace shell logic for OS release with Facter
To generate the correct init script components, install.rake
was shelling out to get a static field from /etc/redhat-release.
This is a fragile technique and Facter uses a much more robust
technique to derive the correct value across numerous RPM-based
platforms. This technique was part of the 1.7.x Facter branch,
so a build-time dependency has been added on Facter 1.7.0 or greater.