(maint) Explicitly set LANG in RPM spec file
LANG is set to 'C' by default, which is problematic because numerous
parts of Ruby have to parse UTF8 files. This will solve a problem with
Fedora 19, which has an umlaut in the release name. If LANG is
set to 'C', the umlaut then leads to unexpected values returned from Facter.