Extract HTTP client functionality to new class
To make the puppetdb terminus work properly in JRuby, we need
to be able to use a different https client (so that we can
avoid trying to use Ruby's SSL code via JRuby).  In order to
achieve this, we need to abstract out the places where
PuppetDB instantiates an HTTP client.

This commit does that.  It moves all references to Puppet's
'HttpPool' class to a new class called 'Puppetdb::HttpClient'.
For now, all this class does is provide an instance of the
HttpPool client that we were already using; but from here,
we can add some more interesting logic to potentially return
other HTTP client implementations.