(PDB-454) - Updating the catalogs endpoint to return environments
This changeset does several things. It changes the catalog endpoint's results
to use the new version 4 wire format. It also lays down some ground work for
a canonical catalog format. Currently there are several different internal representations
of a catalog in the code. This attempts to unify it with our new definition of
wire format versions and enforce it with Schema.