(PDB-567 & PDB-191) Use hash not config_version for report export files
Some users have complained of the inability for the compression library to
translate the configuration version when it contains a forward slash. Not
surprising, I'm sure other parseability and link safety issues will come
from this.

It comes down to the configuration version proving not to be a great
token for uniqueness as the format can be very variable and possibly
not link safe.

This patch changes the configuration version to instead use the a hash
of the combination of state-time and configuration version, and during
anonymization we've stopped using the regexp to obtain any information
(including the hostname) so the name of the file is no longer relevant.

The ticket PDB-567 also suggested an alternative was to upgrade the
compression library to achieve the ability to parse / safely. Since this
is a good idea anyway, I've also upgrade that.

Signed-off-by: Ken Barber <ken@bob.sh>