(PDB-697) Test and document new reports features
Moving reports to the new query engine enabled querying on many new
fields and all of the "standard" operators. This commit adds tests for
each of those fields and add docs around them.

This commit also fixes and issue with testing resources. The previous
code was not running some of the query tests on the new query engine.
The tests at the http layer were going through to the query engine
correctly, but the ones below that layer were not.