(PDB-660) Final refactor for streaming
This patch is the final refactor for streaming:

* It removes the add-limit-clause function because limits are a thing of the past.
* Minute stylistic cleanups, like :refer conversion to :as where applicable.
* Some doc strings have been fixed.
* Removal of limited-query-resource for resources.clj, as it was overlooked.
* More conversion from :use to :require where noticed.

It also moved the old query new/query decision making into the query->sql
function to ensure the tests can use the new query code.

This surfaced a lot of new errors, and fixes to the new query-eng have been
included:

* Handling [not ["=" "property" "foo"]], events was expecting the nils to be returned
  in this case, which is not the normal sql semantics, we fake it now by including
  'is not nil' in our clauses.
* latest-report? was fixed
* Some hyphen/underscore problems fixed with nodes endpoint
* handle new-value old-value serialization

In the process some tests were improved and their verbosity increased somewhat.

Signed-off-by: Ken Barber <ken@bob.sh>